refactor(client): 抽取中心列定位/测量链到 sidebar/use-center-column.ts
Sidebar.tsx 拆分第二块：centerRectRef/centerMeasured/centerColRef/
draggingRef、measureCenter（ref + 直写 DOM，避免动画帧级 re-render）
与整条 locate/ResizeObserver/MutationObserver/1.5s 重试链原样迁出到
src/client/sidebar/use-center-column.ts，注释逐字保留。bottomRef 与
state?.bottomOpen 由调用方传入（bottomRef 在主文件先建再共享，行为与
原先 measureCenter 闭包捕获一致）。

行为不变：hook 每次 render 全量重跑，effect 依赖与值完全相同；唯一
文字差异是三处依赖数组补列了跨 hook 边界后失去 useRef 溯源的稳定
ref 对象（centerColRef/draggingRef，标识恒定，不改变触发时机），
为 react-hooks/exhaustive-deps 零告警所需。
